Bugs item #3610360, was opened at 2013-04-09 00:09 Message generated for change (Tracker Item Submitted) made by paolobenve You can respond by visiting: https://sourceforge.net/tracker/?func=detail&atid=655717&aid=3610360&group_id=110216
Please note that this message will contain a full copy of the comment thread,
including the initial issue submission, for this request,
not just the latest update.
Category: Problem with Italian
Group: None
Status: Open
Resolution: None
Priority: 5
Private: No
Submitted By: Paolo Benvenuto (paolobenve)
Assigned to: Paolo Bianchini (pbianchini)
Summary: number in date

Initial Comment:
In my text I have "sabato 20 Aprile 2013", and the tool reports that 20 has to be changed to the corresponding in letters.

However, in such a case everything is fine: a number after a week day and/or before a month name should be ok.


git-svn-id: https://languagetool.svn.sourceforge.net/svnroot/languagetool/trunk/languagetool@9902 2fe6a88b-1217-0410-b804-ba6ed01314b8